authentication added

+++ routes/web.php
@@ -1,31 +1,29 @@
 <?php
 
 /*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
+  |--------------------------------------------------------------------------
+  | Web Routes
+  |--------------------------------------------------------------------------
+  |
+  | Here is where you can register web routes for your application. These
+  | routes are loaded by the RouteServiceProvider within a group which
+  | contains the "web" middleware group. Now create something great!
+  |
+ */
 
-Route::post('/products', 'ProductsController@store');
-Route::get('/products/create', 'ProductsController@create');
+Route::post('/products', 'ProductsController@store')->middleware('auth');
+Route::get('/products/create', 'ProductsController@create')->middleware('auth');
 Route::get('/products/{product}', 'ProductsController@show');
-Route::delete('/products/{id}', 'ProductsController@destroy');
-Route::patch('/products/{product}', 'ProductsController@update');
+Route::delete('/products/{id}', 'ProductsController@destroy')->middleware('auth');
+Route::patch('/products/{product}', 'ProductsController@update')->middleware('auth');
 Route::get('/products', 'ProductsController@index');
 
-Route::get('/prices/create/{product}', 'PricesController@create');
-Route::post('/prices/{product}', 'PricesController@store');
-Route::patch('/prices/{price}', 'PricesController@update');
-Route::delete('/prices/{price}', 'PricesController@destroy');
+Route::get('/prices/create/{product}', 'PricesController@create')->middleware('auth');
+Route::post('/prices/{product}', 'PricesController@store')->middleware('auth');
+Route::patch('/prices/{price}', 'PricesController@update')->middleware('auth');
+Route::delete('/prices/{price}', 'PricesController@destroy')->middleware('auth');
 
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/', 'ProductsController@index');
 
 Auth::routes();
